requirements.txt
Django
psycopg2-binary
djangorestframework
python-dotenv
Pillow
django-tailwind
